Change default .env to development files
This is a proposal to make local development easier. It will allow
people to start developing without figuring out why the API doesn't
start correctly.

Unfortunately NestJS's env config implementation is broken. Normally
when you have a competing ENV variables set, it will prefer that over
what is inside the file. So to avoid that, when NODE_ENV is in production,
it will only use the machine environment variables.

We also need to put the staging in production mode for this fix to
work.