Move event scheduling into uptime service (#781)
I did this because I'm starting to see an issue with our code base in
that it's not clear who owns what responsibility, and I see a lot of
code fragmentation. I'm not sure what to do about it, though I just feel
like a "NodeUptime" service should be in charge of also creating or
scheduling the creation of the event it's related to.