Merged status calculation for workflow
We had 2 different functions for calculating status of a workflow
result, one for standalone, one for normal. However, the models are
designed such that each section is just optional if the workflow doesn't
specify that. You can easily combine them if you just check if the
section is present, then validate that section.